+++ plugins/src/bh2_example.h
@@ -21,7 +21,7 @@ protected:
 	Sample envelope;
 public:
 	// Change the name and define parameters in constructor
-	bh2_example_effect() : BH2_effect_base(512u), delayLine(aux::millisecToSamples(2502u))
+	bh2_example_effect() : BH2_effect_base(512u), delayLine(size_t(aux::millisecToSamples(2502u)))
 	{
 		// Don't change this //////////////////////////////////////////
 		this->currentParameters = new float[NUMBER_OF_PARAMETERS];   //
@@ -64,7 +64,7 @@ public:
 		float& threshold = parameters[2];
 		float& attack = parameters[3];
 		bool limiterIsOn = parameters[4] > 0.5 ? true : false;
-		size_t attackTimeInSamples = aux::millisecToSamples(attack);
+		size_t attackTimeInSamples = size_t(aux::millisecToSamples(attack));
 		delayLine.setSize(attackTimeInSamples);
 		for (size_t i = 0; i < numberOfSamples; i++)
 		{
